Add methods to fill words with correct letters
Add method for correct letter guesses that fills and displays the
letters of the secret word the match the correct letter guess.

Additionally, seperate and create individual methods for filling
an empty word state (no guesses yet) and displaying the current word
state. Previous combining of the 2 was causing bug.